put success and error cases together
