+++ documentdb_tests/compatibility/tests/system/administration/commands/setParameter/test_setParameter_bson_type_validation.py
@@ -1,4 +1,4 @@
-"""Tests for setParameter BSON type validation.
+"""Tests for setParameter BSON type validation (success cases).
 
 Validates control field acceptance of all BSON types, and type coercion
 behavior for boolean and integer parameter values.
@@ -7,13 +7,12 @@ behavior for boolean and integer parameter values.
 import pytest
 from bson import Decimal128, Int64
 
-from documentdb_tests.framework.assertions import assertFailureCode, assertSuccessPartial
+from documentdb_tests.framework.assertions import assertSuccessPartial
 from documentdb_tests.framework.bson_type_validator import (
     BsonType,
     BsonTypeTestCase,
     generate_bson_acceptance_test_cases,
 )
-from documentdb_tests.framework.error_codes import BAD_VALUE_ERROR
 from documentdb_tests.framework.executor import execute_admin_command
 
 pytestmark = [pytest.mark.admin, pytest.mark.no_parallel]
@@ -44,37 +43,26 @@ def test_setParameter_control_field_bson_type_accepted(collection, bson_type, sa
 @pytest.mark.parametrize(
     "value,desc",
     [
-        (True, "bool True"),
-        (False, "bool False"),
-        (1, "int 1"),
-        (0, "int 0"),
-        (1.0, "double 1.0"),
-        (0.0, "double 0.0"),
-        (Int64(1), "Int64(1)"),
-        (Int64(0), "Int64(0)"),
-        ("true", "string 'true'"),
-        ([True], "array [True]"),
-        ({"a": True}, "document"),
-    ],
-    ids=[
-        "true",
-        "false",
-        "int1",
-        "int0",
-        "double1",
-        "double0",
-        "long1",
-        "long0",
-        "string",
-        "array",
-        "document",
+        pytest.param(True, "bool True", id="true"),
+        pytest.param(False, "bool False", id="false"),
+        pytest.param(1, "int 1", id="int1"),
+        pytest.param(0, "int 0", id="int0"),
+        pytest.param(1.0, "double 1.0", id="double1"),
+        pytest.param(0.0, "double 0.0", id="double0"),
+        pytest.param(Int64(1), "Int64(1)", id="long1"),
+        pytest.param(Int64(0), "Int64(0)", id="long0"),
+        pytest.param("true", "string 'true'", id="string"),
+        pytest.param([True], "array [True]", id="array"),
+        pytest.param({"a": True}, "document", id="document"),
     ],
 )
 def test_setParameter_boolean_coercion_accepted(collection, value, desc):
-    """Test boolean parameter coercion — MongoDB 8.2 coerces many types to boolean."""
+    """Test setParameter boolean parameter coercion."""
     original = execute_admin_command(collection, {"getParameter": 1, "quiet": 1})
     result = execute_admin_command(collection, {"setParameter": 1, "quiet": value})
-    assertSuccessPartial(result, {"ok": 1.0}, msg=f"Boolean param should accept {desc}")
+    assertSuccessPartial(
+        result, {"ok": 1.0}, msg=f"setParameter boolean param should accept {desc}"
+    )
     execute_admin_command(collection, {"setParameter": 1, "quiet": original["quiet"]})
 
 
@@ -82,32 +70,17 @@ def test_setParameter_boolean_coercion_accepted(collection, value, desc):
 @pytest.mark.parametrize(
     "value,desc",
     [
-        (1, "int32"),
-        (Int64(1), "Int64"),
-        (1.0, "whole double"),
-        (Decimal128("1"), "Decimal128 whole"),
-        (True, "bool True"),
+        pytest.param(1, "int32", id="int32"),
+        pytest.param(Int64(1), "Int64", id="long"),
+        pytest.param(1.0, "whole double", id="whole_double"),
+        pytest.param(Decimal128("1"), "Decimal128 whole", id="decimal128_whole"),
+        pytest.param(True, "bool True", id="bool"),
     ],
-    ids=["int32", "long", "whole_double", "decimal128_whole", "bool"],
 )
 def test_setParameter_integer_coercion_accepted(collection, value, desc):
-    """Test integer parameter coercion — whole-number numeric types and bool accepted."""
+    """Test setParameter integer parameter coercion."""
     result = execute_admin_command(collection, {"setParameter": 1, "logLevel": value})
-    assertSuccessPartial(result, {"ok": 1.0}, msg=f"Integer param should accept {desc}")
+    assertSuccessPartial(
+        result, {"ok": 1.0}, msg=f"setParameter integer param should accept {desc}"
+    )
     execute_admin_command(collection, {"setParameter": 1, "logLevel": 0})
-
-
-# Property [Integer Coercion Rejected]: integer-typed params reject non-numeric types.
-@pytest.mark.parametrize(
-    "value,error_code,desc",
-    [
-        ("1", BAD_VALUE_ERROR, "string '1'"),
-        ([1], BAD_VALUE_ERROR, "array [1]"),
-        ({"a": 1}, BAD_VALUE_ERROR, "document"),
-    ],
-    ids=["string", "array", "document"],
-)
-def test_setParameter_integer_coercion_rejected(collection, value, error_code, desc):
-    """Test integer parameter rejects non-numeric types."""
-    result = execute_admin_command(collection, {"setParameter": 1, "logLevel": value})
-    assertFailureCode(result, error_code, msg=f"Integer param should reject {desc}")
